feat: rm hooks, ensure mailchimp email updates
